Se elimino la exportacion de buttonVariants en el componente button
Se movió la definición de buttonVariants a button-variants.ts (nuevo archivo) para evitar advertencias y problemas con Fast Refresh/Hot Reload.
React recomienda que los archivos de componentes solo exporten componentes para que el estado y la recarga en caliente funcionen correctamente.
